struct: Move all from binary-search-tree lib to app
