+++ app/build.gradle.kts
@@ -12,7 +12,6 @@ dependencies {
     testImplementation(kotlin("test"))
     testImplementation(platform("org.junit:junit-bom:5.9.2"))
     testImplementation("org.junit.jupiter:junit-jupiter")
-    implementation(project(":binary-search-trees"))
 }
 
 tasks.test {
