+++ app/src/main/kotlin/app/model/bst/AVLTree.kt
@@ -1,0 +1,9 @@
+package app.model.bst
+
+import app.model.bst.balancer.AVLBalancer
+import app.model.bst.node.AVLTreeNode
+
+class AVLTree<E : Comparable<E>> :
+    BinarySearchTree<E, AVLTreeNode<E>>(
+        balancer = AVLBalancer(),
+    )
