+++ app/src/main/kotlin/app/model/bst/node/AVLTreeNode.kt
@@ -1,0 +1,5 @@
+package app.model.bst.node
+
+interface AVLTreeNode<E : Comparable<E>> : BinTreeNode<E, AVLTreeNode<E>> {
+    var height: Int
+}
