+++ app/src/main/kotlin/app/model/bst/node/BSNode.kt
@@ -1,0 +1,7 @@
+package app.model.bst.node
+
+internal class BSNode<E : Comparable<E>>(
+    override var value: E,
+    override var left: BinSearchTreeNode<E>? = null,
+    override var right: BinSearchTreeNode<E>? = null,
+) : BinSearchTreeNode<E>
