+++ app/src/main/kotlin/app/model/bst/node/BinSearchTreeNode.kt
@@ -1,0 +1,3 @@
+package app.model.bst.node
+
+interface BinSearchTreeNode<E : Comparable<E>> : BinTreeNode<E, BinSearchTreeNode<E>>
