+++ app/src/main/kotlin/app/model/bst/node/BinTreeNode.kt
@@ -1,0 +1,7 @@
+package app.model.bst.node
+
+interface BinTreeNode<E : Comparable<E>, Subtype : BinTreeNode<E, Subtype>> {
+    var value: E
+    var left: Subtype?
+    var right: Subtype?
+}
