+++ app/src/main/kotlin/app/model/bst/node/RBNode.kt
@@ -1,0 +1,9 @@
+package app.model.bst.node
+
+internal class RBNode<E : Comparable<E>>(
+    override var value: E,
+    override var parent: RedBlackTreeNode<E>? = null,
+    override var color: RedBlackTreeNode.Color = RedBlackTreeNode.Color.RED,
+    override var left: RedBlackTreeNode<E>? = null,
+    override var right: RedBlackTreeNode<E>? = null,
+) : RedBlackTreeNode<E>
