--- binary-search-trees/build.gradle.kts
@@ -1,21 +1,0 @@
-plugins {
-    id("org.jetbrains.kotlin.jvm") version "1.8.10"
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation(kotlin("test"))
-    testImplementation(platform("org.junit:junit-bom:5.9.2"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-
-kotlin {
-    jvmToolchain(8)
-}
