--- binary-search-trees/src/main/kotlin/bst/AVLTree.kt
@@ -1,9 +1,0 @@
-package bst
-
-import bst.balancer.AVLBalancer
-import bst.node.AVLTreeNode
-
-class AVLTree<E : Comparable<E>> :
-    BinarySearchTree<E, AVLTreeNode<E>>(
-        balancer = AVLBalancer(),
-    )
