--- binary-search-trees/src/main/kotlin/bst/BSTree.kt
@@ -1,9 +1,0 @@
-package bst
-
-import bst.balancer.BSBalancer
-import bst.node.BinSearchTreeNode
-
-class BSTree<E : Comparable<E>> :
-    BinarySearchTree<E, BinSearchTreeNode<E>>(
-        balancer = BSBalancer(),
-    )
