--- binary-search-trees/src/main/kotlin/bst/RBTree.kt
@@ -1,9 +1,0 @@
-package bst
-
-import bst.balancer.RBBalancer
-import bst.node.RedBlackTreeNode
-
-class RBTree<E : Comparable<E>> :
-    BinarySearchTree<E, RedBlackTreeNode<E>>(
-        balancer = RBBalancer(),
-    )
