--- binary-search-trees/src/main/kotlin/bst/node/AVLTreeNode.kt
@@ -1,5 +1,0 @@
-package bst.node
-
-abstract class AVLTreeNode<E : Comparable<E>> : BinTreeNode<E, AVLTreeNode<E>>() {
-    internal abstract var height: Int
-}
