--- binary-search-trees/src/main/kotlin/bst/node/BinSearchTreeNode.kt
@@ -1,3 +1,0 @@
-package bst.node
-
-abstract class BinSearchTreeNode<E : Comparable<E>> : BinTreeNode<E, BinSearchTreeNode<E>>()
